requirements/development.txt
-r base.txt

# Development tools
django-debug-toolbar>=5.0,<6.0
django-extensions>=3.2,<4.0
